move dataclasses and clusters to separate modules
