+++ test/test_pg_dump.py
@@ -1,6 +1,7 @@
 # Copyright (c) 2020 Aiven, Helsinki, Finland. https://aiven.io/
 
-from aiven_db_migrate.migrate.pgmigrate import PGDatabase, PGMigrate
+from aiven_db_migrate.migrate.models import PGDatabase
+from aiven_db_migrate.migrate.pgmigrate import PGMigrate
 from test.utils import PGRunner, random_string
 from typing import Tuple
