+++ test/test_pg_extensions.py
@@ -1,7 +1,8 @@
 # Copyright (c) 2020 Aiven, Helsinki, Finland. https://aiven.io/
 
 from aiven_db_migrate.migrate.errors import PGMigrateValidationFailedError
-from aiven_db_migrate.migrate.pgmigrate import PGDatabase, PGExtension, PGMigrate
+from aiven_db_migrate.migrate.models import PGDatabase, PGExtension
+from aiven_db_migrate.migrate.pgmigrate import PGMigrate
 from packaging.version import Version
 from test.utils import PGRunner, random_string
 from typing import Tuple
