+++ test/test_pg_migrate.py
@@ -1,7 +1,8 @@
 # Copyright (c) 2020 Aiven, Helsinki, Finland. https://aiven.io/
 
+from aiven_db_migrate.migrate import PGMigrateResult
 from aiven_db_migrate.migrate.errors import PGMigrateValidationFailedError
-from aiven_db_migrate.migrate.pgmigrate import PGMigrate, PGMigrateResult
+from aiven_db_migrate.migrate.pgmigrate import PGMigrate
 from test.utils import PGRunner, random_string, Timer
 from typing import Any, Dict, Optional
